backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
dj-rest-auth>=4.0.0
django-allauth>=0.54.0
python-decouple>=3.8
django-cors-headers>=4.0.0
Pillow>=9.5.0
requests>=2.31.0
psycopg2-binary>=2.9.0
cryptography